requirements.txt
tensorflow>=2.0.0
matplotlib>=3.0.0
numpy>=1.19.0
Pillow>=8.0.0
h5py>=2.10.0
streamlit>=1.20.0
groq
python-dotenv
langchain
langchain-groq
langchain-community
gdown
